move filesystem controller to new world
this cheats a bit and has an instance of FilesystemModel in both the server and the client
