+++ po/POTFILES.in
@@ -1,6 +1,7 @@
 [encoding: UTF-8]
 subiquity/client/client.py
 subiquity/client/controller.py
+subiquity/client/controllers/filesystem.py
 subiquity/client/controllers/__init__.py
 subiquity/client/controllers/keyboard.py
 subiquity/client/controllers/mirror.py
@@ -28,20 +29,20 @@ subiquity/common/api/tests/test_client.py
 subiquity/common/api/tests/test_endtoend.py
 subiquity/common/api/tests/test_server.py
 subiquity/common/errorreport.py
+subiquity/common/filesystem.py
 subiquity/common/__init__.py
 subiquity/common/keyboard.py
 subiquity/common/serialize.py
+subiquity/common/tests/__init__.py
+subiquity/common/tests/test_filesystem.py
 subiquity/common/tests/test_keyboard.py
 subiquity/common/types.py
 subiquity/controller.py
-subiquity/controllers/filesystem.py
 subiquity/controllers/identity.py
 subiquity/controllers/__init__.py
 subiquity/controllers/reboot.py
 subiquity/controllers/snaplist.py
 subiquity/controllers/ssh.py
-subiquity/controllers/tests/__init__.py
-subiquity/controllers/tests/test_filesystem.py
 subiquitycore/async_helpers.py
 subiquitycore/contextlib38.py
 subiquitycore/context.py
@@ -118,6 +119,7 @@ subiquity/models/tests/test_subiquity.py
 subiquity/server/controller.py
 subiquity/server/controllers/cmdlist.py
 subiquity/server/controllers/debconf.py
+subiquity/server/controllers/filesystem.py
 subiquity/server/controllers/__init__.py
 subiquity/server/controllers/install.py
 subiquity/server/controllers/keyboard.py
