+++ subiquity/common/filesystem.py
@@ -1,4 +1,4 @@
-# Copyright 2015 Canonical, Ltd.
+# Copyright 2020 Canonical, Ltd.
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU Affero General Public License as
@@ -13,382 +13,25 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import asyncio
-import json
 import logging
-import os
-import select
 
-import pyudev
-
-from subiquitycore.async_helpers import (
-    run_in_thread,
-    schedule_task,
-    SingleInstanceTask,
-    )
-from subiquitycore.context import with_context
-from subiquitycore.lsb_release import lsb_release
-from subiquitycore.utils import (
-    run_command,
-    )
-
-
-from subiquity.common.errorreport import ErrorReportKind
 from subiquity.common.types import Bootloader
-from subiquity.controller import SubiquityTuiController
 from subiquity.models.filesystem import (
     align_up,
-    DeviceAction,
     dehumanize_size,
+    DeviceAction,
     Partition,
-    raidlevels_by_value,
-    )
-from subiquity.ui.views import (
-    FilesystemView,
-    GuidedDiskSelectionView,
-    )
-from subiquity.ui.views.filesystem.probing import (
-    SlowProbing,
-    ProbingFailed,
     )
 
+log = logging.getLogger('subiquity.common.filesystem')
 
-log = logging.getLogger("subiquitycore.controller.filesystem")
-block_discover_log = logging.getLogger('block-discover')
 
 BIOS_GRUB_SIZE_BYTES = 1 * 1024 * 1024    # 1MiB
 PREP_GRUB_SIZE_BYTES = 8 * 1024 * 1024    # 8MiB
 UEFI_GRUB_SIZE_BYTES = 512 * 1024 * 1024  # 512MiB EFI partition
 
 
-class FilesystemController(SubiquityTuiController):
-
-    autoinstall_key = "storage"
-    autoinstall_schema = {'type': 'object'}  # ...
-    model_name = "filesystem"
-
-    def __init__(self, app):
-        self.ai_data = {}
-        super().__init__(app)
-        self.model.target = app.base_model.target
-        if self.opts.dry_run and self.opts.bootloader:
-            name = self.opts.bootloader.upper()
-            self.model.bootloader = getattr(Bootloader, name)
-        self.answers.setdefault('guided', False)
-        self.answers.setdefault('guided-index', 0)
-        self.answers.setdefault('manual', [])
-        self._monitor = None
-        self._crash_reports = {}
-        self._probe_once_task = SingleInstanceTask(
-            self._probe_once, propagate_errors=False)
-        self._probe_task = SingleInstanceTask(
-            self._probe, propagate_errors=False)
-        if self.model.bootloader == Bootloader.PREP:
-            self.supports_resilient_boot = False
-        else:
-            release = lsb_release()['release']
-            self.supports_resilient_boot = release >= '20.04'
-
-    def load_autoinstall_data(self, data):
-        log.debug("load_autoinstall_data %s", data)
-        if data is None:
-            if not self.interactive():
-                data = {
-                    'layout': {
-                        'name': 'lvm',
-                        },
-                    }
-            else:
-                data = {}
-        log.debug("self.ai_data = %s", data)
-        self.ai_data = data
-
-    @with_context()
-    async def apply_autoinstall_config(self, context=None):
-        self.stop_listening_udev()
-        await self._start_task
-        await self._probe_task.wait()
-        self.convert_autoinstall_config(context=context)
-        if not self.model.is_root_mounted():
-            raise Exception("autoinstall config did not mount root")
-        if self.model.needs_bootloader_partition():
-            raise Exception(
-                "autoinstall config did not create needed bootloader "
-                "partition")
-
-    @with_context(name='probe_once', description='restricted={restricted}')
-    async def _probe_once(self, *, context, restricted):
-        if restricted:
-            probe_types = {'blockdev'}
-            fname = 'probe-data-restricted.json'
-            key = "ProbeDataRestricted"
-        else:
-            probe_types = None
-            fname = 'probe-data.json'
-            key = "ProbeData"
-        storage = await run_in_thread(
-            self.app.prober.get_storage, probe_types)
-        fpath = os.path.join(self.app.block_log_dir, fname)
-        with open(fpath, 'w') as fp:
-            json.dump(storage, fp, indent=4)
-        self.app.note_file_for_apport(key, fpath)
-        self.model.load_probe_data(storage)
-
-    @with_context()
-    async def _probe(self, *, context=None):
-        async with self.app.install_lock_file.shared():
-            self._crash_reports = {}
-            if isinstance(self.ui.body, ProbingFailed):
-                self.ui.set_body(SlowProbing(self))
-                schedule_task(self._wait_for_probing())
-            for (restricted, kind) in [
-                    (False, ErrorReportKind.BLOCK_PROBE_FAIL),
-                    (True,  ErrorReportKind.DISK_PROBE_FAIL),
-                    ]:
-                try:
-                    await self._probe_once_task.start(
-                        context=context, restricted=restricted)
-                    # We wait on the task directly here, not
-                    # self._probe_once_task.wait as if _probe_once_task
-                    # gets cancelled, we should be cancelled too.
-                    await asyncio.wait_for(self._probe_once_task.task, 15.0)
-                except asyncio.CancelledError:
-                    # asyncio.CancelledError is a subclass of Exception in
-                    # Python 3.6 (sadface)
-                    raise
-                except Exception:
-                    block_discover_log.exception(
-                        "block probing failed restricted=%s", restricted)
-                    report = self.app.make_apport_report(
-                        kind, "block probing", interrupt=False)
-                    if report is not None:
-                        self._crash_reports[restricted] = report
-                    continue
-                break
-
-    @with_context()
-    def convert_autoinstall_config(self, context=None):
-        log.debug("self.ai_data = %s", self.ai_data)
-        if 'layout' in self.ai_data:
-            layout = self.ai_data['layout']
-            meth = getattr(self, "guided_" + layout['name'])
-            disk = self.model.disk_for_match(
-                self.model.all_disks(),
-                layout.get("match", {'size': 'largest'}))
-            meth(disk)
-        elif 'config' in self.ai_data:
-            self.model.apply_autoinstall_config(self.ai_data['config'])
-            self.model.grub = self.ai_data.get('grub', {})
-            self.model.swap = self.ai_data.get('swap')
-
-    def start(self):
-        self._start_task = schedule_task(self._start())
-
-    async def _start(self):
-        context = pyudev.Context()
-        self._monitor = pyudev.Monitor.from_netlink(context)
-        self._monitor.filter_by(subsystem='block')
-        self._monitor.enable_receiving()
-        self.start_listening_udev()
-        await self._probe_task.start()
-
-    def start_listening_udev(self):
-        loop = asyncio.get_event_loop()
-        loop.add_reader(self._monitor.fileno(), self._udev_event)
-
-    def stop_listening_udev(self):
-        loop = asyncio.get_event_loop()
-        loop.remove_reader(self._monitor.fileno())
-
-    def _udev_event(self):
-        cp = run_command(['udevadm', 'settle', '-t', '0'])
-        if cp.returncode != 0:
-            log.debug("waiting 0.1 to let udev event queue settle")
-            self.stop_listening_udev()
-            loop = asyncio.get_event_loop()
-            loop.call_later(0.1, self.start_listening_udev)
-            return
-        # Drain the udev events in the queue -- if we stopped listening to
-        # allow udev to settle, it's good bet there is more than one event to
-        # process and we don't want to kick off a full block probe for each
-        # one.  It's a touch unfortunate that pyudev doesn't have a
-        # non-blocking read so we resort to select().
-        while select.select([self._monitor.fileno()], [], [], 0)[0]:
-            action, dev = self._monitor.receive_device()
-            log.debug("_udev_event %s %s", action, dev)
-        self._probe_task.start_sync()
-
-    async def _wait_for_probing(self):
-        await self._start_task
-        await self._probe_task.wait()
-        if isinstance(self.ui.body, SlowProbing):
-            self.ui.set_body(self.make_ui())
-
-    def make_ui(self):
-        if self._probe_task.task is None or not self._probe_task.task.done():
-            schedule_task(self._wait_for_probing())
-            return SlowProbing(self)
-        elif True in self._crash_reports:
-            pr = self._crash_reports[True]
-            if pr is not None:
-                self.app.show_error_report(pr.ref())
-            return ProbingFailed(self)
-        else:
-            # Once we've shown the filesystem UI, we stop listening for udev
-            # events as merging system changes with configuration the user has
-            # performed would be tricky.  Possibly worth doing though! Just
-            # not today.
-            self.convert_autoinstall_config()
-            self.stop_listening_udev()
-            pr = self._crash_reports.get(False)
-            if pr is not None:
-                self.app.show_error_report(pr.ref())
-            if self.answers:
-                self.app.aio_loop.call_soon(self._start_answers)
-            return GuidedDiskSelectionView(self)
-
-    def _start_answers(self):
-        if self.answers['guided']:
-            disk = self.model.all_disks()[self.answers['guided-index']]
-            method = self.answers.get('guided-method')
-            self.ui.body.form.guided_choice.value = {
-                'disk': disk,
-                'use_lvm': method == "lvm",
-                }
-            self.ui.body.done(self.ui.body.form)
-        elif self.answers['manual']:
-            self.manual()
-
-    def run_answers(self):
-        # Handled above as we only want to run answers when probing
-        # completes.
-        pass
-
-    def _action_get(self, id):
-        dev_spec = id[0].split()
-        dev = None
-        if dev_spec[0] == "disk":
-            if dev_spec[1] == "index":
-                dev = self.model.all_disks()[int(dev_spec[2])]
-            elif dev_spec[1] == "serial":
-                dev = self.model._one(type='disk', serial=dev_spec[2])
-        elif dev_spec[0] == "raid":
-            if dev_spec[1] == "name":
-                for r in self.model.all_raids():
-                    if r.name == dev_spec[2]:
-                        dev = r
-                        break
-        elif dev_spec[0] == "volgroup":
-            if dev_spec[1] == "name":
-                for r in self.model.all_volgroups():
-                    if r.name == dev_spec[2]:
-                        dev = r
-                        break
-        if dev is None:
-            raise Exception("could not resolve {}".format(id))
-        if len(id) > 1:
-            part, index = id[1].split()
-            if part == "part":
-                return dev.partitions()[int(index)]
-        else:
-            return dev
-        raise Exception("could not resolve {}".format(id))
-
-    def _action_clean_devices_raid(self, devices):
-        r = {
-            self._action_get(d): v
-            for d, v in zip(devices[::2], devices[1::2])
-            }
-        for d in r:
-            assert d.ok_for_raid
-        return r
-
-    def _action_clean_devices_vg(self, devices):
-        r = {self._action_get(d): 'active' for d in devices}
-        for d in r:
-            assert d.ok_for_lvm_vg
-        return r
-
-    def _action_clean_level(self, level):
-        return raidlevels_by_value[level]
-
-    async def _answers_action(self, action):
-        from subiquitycore.ui.stretchy import StretchyOverlay
-        from subiquity.ui.views.filesystem.delete import ConfirmDeleteStretchy
-        log.debug("_answers_action %r", action)
-        if 'obj' in action:
-            obj = self._action_get(action['obj'])
-            action_name = action['action']
-            if action_name == "MAKE_BOOT":
-                action_name = "TOGGLE_BOOT"
-            meth = getattr(
-                self.ui.body.avail_list,
-                "_{}_{}".format(obj.type, action_name))
-            meth(obj)
-            yield
-            body = self.ui.body._w
-            if not isinstance(body, StretchyOverlay):
-                return
-            if isinstance(body.stretchy, ConfirmDeleteStretchy):
-                if action.get("submit", True):
-                    body.stretchy.done()
-            else:
-                async for _ in self._enter_form_data(
-                        body.stretchy.form,
-                        action['data'],
-                        action.get("submit", True)):
-                    pass
-        elif action['action'] == 'create-raid':
-            self.ui.body.create_raid()
-            yield
-            body = self.ui.body._w
-            async for _ in self._enter_form_data(
-                    body.stretchy.form,
-                    action['data'],
-                    action.get("submit", True),
-                    clean_suffix='raid'):
-                pass
-        elif action['action'] == 'create-vg':
-            self.ui.body.create_vg()
-            yield
-            body = self.ui.body._w
-            async for _ in self._enter_form_data(
-                    body.stretchy.form,
-                    action['data'],
-                    action.get("submit", True),
-                    clean_suffix='vg'):
-                pass
-        elif action['action'] == 'done':
-            if not self.ui.body.done.enabled:
-                raise Exception("answers did not provide complete fs config")
-            self.finish(self.app.confirm_install())
-        else:
-            raise Exception("could not process action {}".format(action))
-
-    def manual(self):
-        self.ui.set_body(FilesystemView(self.model, self))
-        if self.answers['guided']:
-            self.finish(self.app.confirm_install())
-        if self.answers['manual']:
-            self.app.aio_loop.create_task(
-                self._run_actions(self.answers['manual']))
-            self.answers['manual'] = []
-
-    def guided(self):
-        self.ui.set_body(GuidedDiskSelectionView(self))
-
-    def reset(self):
-        log.info("Resetting Filesystem model")
-        self.model.reset()
-        self.manual()
-
-    def cancel(self):
-        self.app.prev_screen()
-
-    def finish(self, coro=None):
-        log.debug("FilesystemController.finish next_screen")
-        self.configured()
-        self.app.next_screen(coro)
+class FilesystemManipulator:
 
     def create_mount(self, fs, spec):
         if spec.get('mount') is None:
@@ -723,6 +366,7 @@ class FilesystemController(SubiquityTuiController):
                 part_size = PREP_GRUB_SIZE_BYTES
             elif bootloader == Bootloader.BIOS:
                 part_size = BIOS_GRUB_SIZE_BYTES
+            log.debug("bootloader %s", bootloader)
             if part_size > new_boot_disk.free_for_partitions:
                 largest_part = max(
                     new_boot_disk.partitions(), key=lambda p: p.size)
@@ -781,12 +425,3 @@ class FilesystemController(SubiquityTuiController):
                 fstype="ext4",
                 mount="/",
                 ))
-
-    def make_autoinstall(self):
-        rendered = self.model.render()
-        r = {
-            'config': rendered['storage']['config']
-            }
-        if 'swap' in rendered:
-            r['swap'] = rendered['swap']
-        return r
