+++ subiquity/controllers/__init__.py
@@ -13,14 +13,12 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from .filesystem import FilesystemController
 from .identity import IdentityController
 from .reboot import RebootController
 from .snaplist import SnapListController
 from .ssh import SSHController
 
 __all__ = [
-    'FilesystemController',
     'IdentityController',
     'RebootController',
     'SnapListController',
