+++ subiquity/ui/views/filesystem/tests/test_lvm.py
@@ -21,7 +21,7 @@ import urwid
 from subiquitycore.testing import view_helpers
 from subiquitycore.view import BaseView
 
-from subiquity.controllers.filesystem import FilesystemController
+from subiquity.client.controllers.filesystem import FilesystemController
 from subiquity.ui.views.filesystem.lvm import VolGroupStretchy
 from subiquity.ui.views.filesystem.tests.test_partition import (
     make_model_and_disk,
