+++ subiquity/ui/views/filesystem/tests/test_raid.py
@@ -6,7 +6,7 @@ import urwid
 from subiquitycore.testing import view_helpers
 from subiquitycore.view import BaseView
 
-from subiquity.controllers.filesystem import FilesystemController
+from subiquity.client.controllers.filesystem import FilesystemController
 from subiquity.models.filesystem import (
     raidlevels_by_value,
     )
